+++ core-ng/src/main/java/core/framework/impl/web/session/SessionImpl.java
@@ -17,6 +17,7 @@ public class SessionImpl implements Session {
     final Set<String> changedFields = new HashSet<>();
     String id;
     boolean invalidated;
+    boolean saved;
 
     @Override
     public Optional<String> get(String key) {