+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 ## Change log
-### 6.10.8 (1/2/2019 - )
+### 6.10.8 (1/2/2019 - 1/4/2019)
+* executor: use "TASK_REJECTED" error code when rejecting task during shutdown
+* session: always try to save session even if on exception flow, in case of invalidate session or pass generated session id